requirements.txt
pyheif
exifread
Pillow